+++ app/src/main/java/org/coepi/android/MainActivity.kt
@@ -2,6 +2,7 @@ package org.coepi.android
 
 import android.app.AlarmManager
 import android.app.PendingIntent
+import android.app.PendingIntent.FLAG_UPDATE_CURRENT
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
@@ -29,8 +30,20 @@ import org.coepi.android.ui.notifications.ReminderAlarmHandler
 import org.coepi.android.ui.onboarding.OnboardingShower
 import org.koin.android.ext.android.inject
 import java.util.Calendar
+import java.util.Calendar.DATE
+import java.util.Calendar.HOUR_OF_DAY
+import java.util.Calendar.MINUTE
+import java.util.Calendar.MONTH
+import java.util.Calendar.YEAR
 
 class MainActivity : AppCompatActivity() {
+
+    companion object {
+        private const val NUMBER_OF_DAYS_TO_SET_THE_REMINDER_ALARM_FOR = 14
+        private const val ALARM_TRIGGER_HOURS = 18
+        private const val ALARM_TRIGGER_MINUTES = 0
+    }
+
     private val rootNav: RootNavigation by inject()
     private val onboardingShower: OnboardingShower by inject()
     private val bleInitializer: BleInitializer by inject()
@@ -55,7 +68,7 @@ class MainActivity : AppCompatActivity() {
         setTheme(AppTheme)
         super.onCreate(savedInstanceState)
         setContentView(activity_main)
-        setReminderAlarmsFor(14)
+        setReminderAlarmsFor(NUMBER_OF_DAYS_TO_SET_THE_REMINDER_ALARM_FOR)
         observeRootNavigation()
         observeUINotifier()
         observeActivityFinisher()
@@ -72,9 +85,10 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun setReminderAlarmsFor(days: Int) {
-        val calendar = Calendar.getInstance()
-        calendar.set(Calendar.HOUR_OF_DAY, 18)
-        calendar.set(Calendar.MINUTE, 0)
+        val calendar = Calendar.getInstance().apply {
+            set(HOUR_OF_DAY, ALARM_TRIGGER_HOURS)
+            set(MINUTE, ALARM_TRIGGER_MINUTES)
+        }
 
         val alarmManager = getSystemService(Context.ALARM_SERVICE) as AlarmManager
 
@@ -151,19 +165,11 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
-fun Calendar.getIdentifierFromDate() : Int {
-    val year = this.get(Calendar.YEAR)
-    val month = this.get(Calendar.MONTH)
-    val date = this.get(Calendar.DATE)
-    val notifIdentifier: String = String.format("%04d%02d%02d", year, month, date)
-    log.d("[Reminder] notifIdentifier: $notifIdentifier")
-    return notifIdentifier.toInt()  ?: 0
-}
+fun Calendar.getIdentifierFromDate() = String.format("%04d%02d%02d", get(YEAR), get(MONTH), get(DATE)).toInt()
 
 fun intentWithRequestCode(code: Int, context: Context) : PendingIntent{
     log.d("[Reminder] scheduling: $code")
     return Intent(context, ReminderAlarmHandler::class.java).let { intent ->
-        intent.putExtra("code", code)
-        PendingIntent.getBroadcast(context, code, intent, 0)
+        PendingIntent.getBroadcast(context, code, intent, FLAG_UPDATE_CURRENT)
     }
 }
